move wasm bindings
