+++ ferveo/Cargo.toml
@@ -3,12 +3,15 @@ name = "ferveo"
 version = "0.1.1"
 edition = "2018"
 license = "GPL-3.0"
-repository = "https://gitlab.com/anoma/ferveo"
+repository = "https://gitlab.com/nucypher/ferveo"
 readme = "README.md"
-description = "An implementation of a DKG protocol for front-running protection on public blockchains."
+description = "An implementation of a DKG protocol for threshold decryption."
 keywords = ["DKG"]
 categories = ["cryptography"]
 authors = ["Heliax AG <hello@heliax.dev>"]
+
+[lib]
+bench = false
 
 [dependencies]
 ark-bls12-381 = "0.4"
@@ -32,19 +35,33 @@ subproductdomain = { path = "../subproductdomain" }
 thiserror = "1.0"
 zeroize = { version = "1.6.0", default-features = false, features = ["derive"] }
 
+# Shared by Python and WASM bindings
+derive_more = { version = "0.99", default-features = false, features = ["from", "as_ref"], optional = true }
+
+# Python bindings
+pyo3 = { version = "0.18.2", features = ["macros"], optional = true }
+generic-array = { version = "0.14.7", optional = true }
+
+# WASM bindings
+console_error_panic_hook = { version = "0.1.7", optional = true }
+getrandom = { version = "0.2", features = ["js"], optional = true }
+js-sys = { version = "0.3.61", optional = true }
+wasm-bindgen = { version = "0.2.74", optional = true }
+wasm-bindgen-derive = { version = "0.1.1", optional = true }
+
 [dev-dependencies]
 criterion = "0.3" # supports pprof, # TODO: Figure out if/how we can update to 0.4
 digest = { version = "0.10.0", features = ["alloc"] }
 pprof = { version = "0.6", features = ["flamegraph", "criterion"] }
 
-[package.metadata.cargo-machete]
-ignored = ["ark-serialize"]
-
-[lib]
-bench = false
+# WASM bindings
+console_error_panic_hook = "0.1.7"
+serde = { version = "1.0", features = ["derive"] }
+wasm-bindgen = { version = "0.2.74", features = ["serde-serialize"] }
 
 [features]
-test-common = []
+bindings-python = ["pyo3", "derive_more", "generic-array"]
+bindings-wasm = ["console_error_panic_hook", "derive_more", "getrandom", "js-sys", "wasm-bindgen", "wasm-bindgen-derive"]
 
 [[example]]
 name = "bench_primitives_size"
@@ -59,3 +76,7 @@ path = "examples/bench_primitives_size.rs"
 name = "benchmarks"
 path = "benches/bench_main.rs"
 harness = false
+
+[package.metadata.cargo-machete]
+ignored = ["ark-serialize", "getrandom"]
+
